+++ CHANGELOG.md
@@ -21,6 +21,8 @@ within it
 - Fixed incorrect songs being displayed when adding albums to the end of the queue
 - Fixed freezing occuring when scrolling through large music libraries
 - Fixed app not responding once music loading completes for large libraries
+- Fixed crash when the last song of the queue gets removed while playing
+- Fixed playback UI not re-appearing after playback ends
 
 #### What's Changed
 - Android Lollipop and Marshmallow support have been dropped